Changes pr, move localstorage and formatting
