+++ React/tic-tac/src/redux/store.js
@@ -2,8 +2,7 @@ import { combineReducers, createStore, applyMiddleware, compose } from 'redux';
 import { createBrowserHistory } from 'history';
 import { connectRouter, routerMiddleware } from 'connected-react-router';
 import thunk from 'redux-thunk';
-
-import { loadState, saveState } from '../localStorage';
+import { loadState, saveState } from '@services/localStorageService';
 
 import formReducer from './form/reducer';
 import reducerAuth from './auth/reducer';
